requirements.txt
django
djangorestframework
Pygments
django-cors-headers
django-oauth-toolkit
django-filter
python-decouple
pillow
mysqlclient
zeep
gevent
gunicorn
django-cacheops
django-prometheus
xlwt
django-jalali-date
django-admin-rangefilter
twentytab-totalsum-admin
sentry-sdk
pyyaml
django-user-agents-last
